Implement React Router
Now we use react router to traverse through the pokedex. This has enabled the selection of any pokemon on the sidebar, and allowed skipping to any id you may already know in the address bar.

The API url is now a prop passed from the main App file to get rid of pulling any information from separate components.

Additionally, ID is no longer handled through the Pokedex state, and is now a prop to allow the router to provide the ID based on url parameters.

Finally, HandleLoadNext and HandleLoadPrev now simply return the next or previous numbers in the order respectively. This is used in a new supplementary function that determines what to do based on keypress.